packages/content-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "content-pipeline"
version = "1.0.0"
description = "Sistema de producao de conteudo via API para Manager Grupo — B2B Healthcare"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}

dependencies = [
    "google-generativeai>=0.8.0",
    "Pillow>=10.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "slowapi>=0.1.9",
    "supabase>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[project.scripts]
content-pipeline = "content_pipeline.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
